resources/yummy-platform/yummy-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
redis[asyncio]==5.1.1
httpx==0.27.2
websockets==13.1
anthropic==0.34.2
openai==1.51.0
google-generativeai==0.8.2
langchain==0.3.1
langchain-community==0.3.1
langchain-anthropic==0.2.1
llama-index==0.11.17
pgvector==0.3.2
celery[redis]==5.4.0
python-dotenv==1.0.1
structlog==24.4.0
prometheus-client==0.21.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
